Float sum

+++ 07_Booleans/charwithcfloat.c
@@ -4,10 +4,10 @@
 
         int i = 17;
         char c = 'c'; /* ASCII VALUE IS 99 */
-        int sum;
+        float sum;
 
         sum = i + c;
-        printf("Value of sum : %d\n", sum);
+        printf("Value of sum : %f\n", sum);
 
         return 0;
